+++ app/src/main/java/cityfreqs/com/pilfershushjammer/ui/InspectorFragment.java
@@ -1,6 +1,7 @@
 package cityfreqs.com.pilfershushjammer.ui;
 
 import android.app.AlertDialog;
+import android.app.Dialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.graphics.Color;
@@ -29,8 +30,6 @@ import cityfreqs.com.pilfershushjammer.utilities.BackgroundChecker;
 
 
 public class InspectorFragment extends Fragment implements InspectorAdapter.RecyclerViewClickListener{
-    //TODO make this a RecyclerView with CardViews
-
     private static final String TAG = "PilferShush_Jammer-INSP";
     private boolean DEBUG;
 
@@ -41,6 +40,8 @@ public class InspectorFragment extends Fragment implements InspectorAdapter.Recy
 
     private ImageButton appInspectButton;
     private ImageButton sdkListButton;
+    private Dialog appInfoDialog;
+    private AlertDialog.Builder dialogBuilder;
 
     protected RecyclerView recyclerView;
     protected RecyclerView.LayoutManager layoutManager;
@@ -106,18 +107,22 @@ public class InspectorFragment extends Fragment implements InspectorAdapter.Recy
         populateElements();
         // init recyclerView
         initDataset();
-        //InspectorAdapter.RecyclerViewClickListener clickListener = null;
         inspectorAdapter = new InspectorAdapter(mDataset);
-        inspectorAdapter.setClickListener(this); // Bind the listener
+        inspectorAdapter.setClickListener(this); // bind the listener
         recyclerView.setAdapter(inspectorAdapter);
         return view;
+    }
+
+    public void onDestroyView() {
+        super.onDestroyView();
+        Log.d(TAG, "destroyView");
+        inspectorAdapter.setClickListener(null); // remove the listener
     }
 
     // clickListener listening...
     @Override
     public void onClick(View view, int position) {
-        Log.d(TAG, "ClickListener position: " + position);
-        setAppInfoDialog(backgroundChecker.getOverrideScanAppEntry(position));
+        displayAppInfoDialog(backgroundChecker.getOverrideScanAppEntry(position));
     }
 
     /**********************************************************************************************/
@@ -136,7 +141,7 @@ public class InspectorFragment extends Fragment implements InspectorAdapter.Recy
         appInspectButton.setOnClickListener(new View.OnClickListener() {
             public void onClick(View v) {
                 // Code here executes on main thread after user presses button
-                userAppCheckDialog();
+                userAppSummary();
             }
         });
 
@@ -147,7 +152,6 @@ public class InspectorFragment extends Fragment implements InspectorAdapter.Recy
             }
         });
 
-        //TODO
         if (backgroundChecker.initChecker(context.getPackageManager())) {
             backgroundChecker.runChecker();
             backgroundChecker.checkAudioBeaconApps();
@@ -156,8 +160,7 @@ public class InspectorFragment extends Fragment implements InspectorAdapter.Recy
             // has a problem getting SDK names from file, null or 0 length
             entryLogger(getResources().getString(R.string.userapp_scan_14), true);
         }
-
-        introText();
+        entryLogger(getResources().getString(R.string.intro_6_3) + "\n", false);
 
         // report
         int audioNum = backgroundChecker.getUserRecordNumApps();
@@ -172,35 +175,41 @@ public class InspectorFragment extends Fragment implements InspectorAdapter.Recy
         } else {
             entryLogger(getResources().getString(R.string.userapp_scan_11) + "\n", false);
         }
-    }
 
-    private void introText() {
-        // simple and understandable statements about app usage.
-        entryLogger(getResources().getString(R.string.intro_6_1) + "\n", false);
-        entryLogger(getResources().getString(R.string.intro_6_2) + "\n", false);
-        entryLogger(getResources().getString(R.string.intro_6_3) + "\n", false);
+        entryLogger("Selectable installed user app summary list below:" , false);
     }
-
 
     /**********************************************************************************************/
 
+    // change to dialog
     private void userAppSummary() {
         if (backgroundChecker != null) {
-            entryLogger("\n--------------------------------------\n", false);
-            entryLogger(getResources().getString(R.string.userapp_scan_13), true);
-            entryLogger(getResources().getString(R.string.userapp_scan_intro_1), false);
-            entryLogger(getResources().getString(R.string.userapp_scan_intro_2) + "\n", false);
-            entryLogger(getResources().getString(R.string.userapp_scan_intro_3), false);
-            entryLogger(getResources().getString(R.string.userapp_scan_intro_4), false);
-            entryLogger(getResources().getString(R.string.userapp_scan_intro_5), false);
-            entryLogger(getResources().getString(R.string.userapp_scan_intro_6), false);
-            entryLogger(getResources().getString(R.string.userapp_scan_intro_7), false);
-            entryLogger(getResources().getString(R.string.userapp_scan_intro_8), false);
-            entryLogger(getResources().getString(R.string.userapp_scan_intro_9), false);
-            entryLogger(getResources().getString(R.string.userapp_scan_intro_10), false);
-            entryLogger(getResources().getString(R.string.userapp_scan_intro_11), false);
-            entryLogger(getResources().getString(R.string.userapp_scan_intro_12) + "\n", false);
-            audioAppEntryLog();
+            dialogBuilder = new AlertDialog.Builder(context);
+            dialogBuilder.setTitle(getResources().getString(R.string.userapp_scan_13));
+            String aboutScan = getResources().getString(R.string.userapp_scan_intro_1)
+                    + "\n\n" + getResources().getString(R.string.userapp_scan_intro_2)
+                    + "\n\n" + getResources().getString(R.string.userapp_scan_intro_3)
+                    + "\n\n" + getResources().getString(R.string.userapp_scan_intro_4)
+                    + "\n\n" + getResources().getString(R.string.userapp_scan_intro_5)
+                    + "\n\n" + getResources().getString(R.string.userapp_scan_intro_6)
+                    + "\n\n" + getResources().getString(R.string.userapp_scan_intro_7)
+                    + "\n\n" + getResources().getString(R.string.userapp_scan_intro_8)
+                    + "\n\n" + getResources().getString(R.string.userapp_scan_intro_9)
+                    + "\n\n" + getResources().getString(R.string.userapp_scan_intro_10)
+                    + "\n\n" + getResources().getString(R.string.userapp_scan_intro_11)
+                    + "\n\n" + getResources().getString(R.string.userapp_scan_intro_12);
+
+            dialogBuilder.setMessage(aboutScan);
+            dialogBuilder.setCancelable(true);
+            dialogBuilder
+                    .setPositiveButton(R.string.dialog_button_okay, new DialogInterface.OnClickListener() {
+                        @Override
+                        public void onClick(DialogInterface dialog, int id) {
+                            Log.d(TAG, "Dismiss sdk list popup");
+                        }
+                    });
+            appInfoDialog = dialogBuilder.create();
+            appInfoDialog.show();
         }
         else {
             if (DEBUG) Log.d(TAG, "backgroundChecker is NULL at userApp summary.");
@@ -208,119 +217,45 @@ public class InspectorFragment extends Fragment implements InspectorAdapter.Recy
         }
     }
 
-    //TODO make each a clickable pop up with full report
-    // needs to be a listview, major rewrite of Inspector
     private void initDataset() {
-        // is zero here
-        int datasetSize = backgroundChecker.getNumberAppEntries();
-        Log.d(TAG, "Dataset size: " + datasetSize);
-        mDataset = new String[datasetSize];
-
-        ArrayList<AppEntry> appEntries = backgroundChecker.getAppEntries();
-        if (appEntries.size() > 0) {
-            int i = 0;
-            for (AppEntry appEntry : appEntries) {
-                mDataset[i] = appEntry.basicEntryPrint();
-                i++;
-            }
-        }
-    }
-
-    private void audioAppEntryLog() {
-        ArrayList<AppEntry> appEntries = backgroundChecker.getAppEntries();
-
-        if (appEntries.size() > 0) {
-            for (AppEntry appEntry : appEntries) {
-                entryLogger(appEntry.entryPrint(), appEntry.checkForCaution());
-            }
-        }
-        else {
-            entryLogger(context.getResources().getString(R.string.userapp_scan_12), false);
-        }
-    }
-    // dialog class?
-    private void setAppInfoDialog(AppEntry appEntry) {
-        AlertDialog.Builder dialogBuilder = new AlertDialog.Builder(context);
-        dialogBuilder.setTitle(appEntry.getActivityName());
-        String appEntryString = appEntry.entryPrint();
-        if (appEntry.getServicesNum() > 0)
-            // better method
-            appEntryString += "\nServices List:\n" + appEntry.printServiceNames();
-        if (appEntry.getReceiversNum() > 0)
-            appEntryString += "\nReceivers List:\n" + appEntry.printReceiverNames();
-        dialogBuilder.setMessage(appEntryString);
-        dialogBuilder.setCancelable(false);
-        dialogBuilder
-             .setPositiveButton(R.string.dialog_button_okay, new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(DialogInterface dialog, int id) {
-                    Log.d(TAG, "Dismiss app info popup");
-                }
-        });
-        AlertDialog appInfoDialog = dialogBuilder.create();
-        appInfoDialog.show();
-    }
-
-    //TODO make whole thing dialogs
-    private void userAppCheckDialog() {
         if (backgroundChecker != null) {
-            String[] appNames = backgroundChecker.getOverrideScanAppNames();
+            int datasetSize = backgroundChecker.getNumberAppEntries();
+            Log.d(TAG, "Dataset size: " + datasetSize);
+            mDataset = new String[datasetSize];
 
-            if (appNames != null && appNames.length > 0) {
-                AlertDialog.Builder dialogBuilder = new AlertDialog.Builder(context);
-                dialogBuilder.setItems(appNames, new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialogInterface, int which) {
-                        // index position of clicked app name
-                        listAppOverrideScanDetails(which);
-                    }
-                });
-                dialogBuilder.setTitle(R.string.dialog_userapps);
-                AlertDialog alertDialog = dialogBuilder.create();
-                alertDialog.show();
-            }
-            else {
-                entryLogger(getResources().getString(R.string.userapp_scan_4), true);
+            ArrayList<AppEntry> appEntries = backgroundChecker.getAppEntries();
+            if (appEntries.size() > 0) {
+                int i = 0;
+                for (AppEntry appEntry : appEntries) {
+                    mDataset[i] = appEntry.basicEntryPrint();
+                    i++;
+                }
             }
         }
         else {
-            if (DEBUG) Log.d(TAG, "backgroundChecker is NULL at userApp check.");
+            if (DEBUG) Log.d(TAG, "backgroundChecker is NULL at initDataset.");
             entryLogger("Background Checker not initialised.", true);
         }
     }
 
-    private void listAppOverrideScanDetails(int selectedIndex) {
-        // already checked for null
-        // check for receivers too?
-        entryLogger("\n" + getResources().getString(R.string.userapp_scan_5)
-                + backgroundChecker.getOverrideScanAppEntry(selectedIndex).getActivityName()
-                + ": " + backgroundChecker.getOverrideScanAppEntry(selectedIndex).getServicesNum(), true);
+    private void displayAppInfoDialog(AppEntry appEntry) {
+        AppEntryDialog appEntryDialog = new AppEntryDialog();
+        // set appEntry basic info with styling
+        appEntryDialog.initAppEntryDialog(context,
+                this.getLayoutInflater(),
+                appEntry);
 
-        if (backgroundChecker.getOverrideScanAppEntry(selectedIndex).getServicesNum() > 0) {
-            logAppEntryInfo(backgroundChecker.getOverrideScanAppEntry(selectedIndex).getServiceNames());
-        }
 
-        entryLogger(getResources().getString(R.string.userapp_scan_6)
-                + backgroundChecker.getOverrideScanAppEntry(selectedIndex).getActivityName()
-                + ": " + backgroundChecker.getOverrideScanAppEntry(selectedIndex).getReceiversNum(), true);
-
-        if (backgroundChecker.getOverrideScanAppEntry(selectedIndex).getReceiversNum() > 0) {
-            logAppEntryInfo(backgroundChecker.getOverrideScanAppEntry(selectedIndex).getReceiverNames());
-        }
+        // build and display
+        appInfoDialog = appEntryDialog.buildDialog();
+        appInfoDialog.show();
     }
 
-    private void logAppEntryInfo(String[] appEntryInfoList) {
-        entryLogger("\n" + getResources().getString(R.string.userapp_scan_7) + "\n", false);
-        for (String entryInfo : appEntryInfoList) {
-            entryLogger(entryInfo + "\n", false);
-        }
-    }
-
-    //TODO make a dialog
     private void displaySdkList() {
         // current list (in /raw) of NUHF/ACR SDK package names
         if (backgroundChecker != null) {
             //String sdkNames = backgroundChecker.displayAudioSdkNames();
-            AlertDialog.Builder dialogBuilder = new AlertDialog.Builder(context);
+            dialogBuilder = new AlertDialog.Builder(context);
             dialogBuilder.setTitle(getResources().getString(R.string.sdk_names_list));
             dialogBuilder.setMessage(backgroundChecker.displayAudioSdkNames());
             dialogBuilder.setCancelable(true);
@@ -331,11 +266,11 @@ public class InspectorFragment extends Fragment implements InspectorAdapter.Recy
                             Log.d(TAG, "Dismiss sdk list popup");
                         }
                     });
-            AlertDialog alertDialog = dialogBuilder.create();
-            alertDialog.show();
+            appInfoDialog = dialogBuilder.create();
+            appInfoDialog.show();
         }
         else {
-            if (DEBUG) Log.d(TAG, "backgroundChecker is NULL.");
+            if (DEBUG) Log.d(TAG, "backgroundChecker is NULL at displaySdkList.");
             entryLogger("Background Checker not initialised.", true);
         }
     }